requirements.txt
bidict==0.22.1
click==8.1.3
colorama==0.4.6
eventlet==0.30.2
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.2
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==0.21.1
python-engineio==4.3.4
python-socketio==5.7.2
six==1.16.0
Werkzeug==2.2.2
firebase-admin==6.0.1
flask-sqlalchemy==3.0.3
psycopg2==2.9.5